Add files via upload
Covid-19, a highly contagious disease with its epicentre in Wuhan, China, caused by coronavirus which is an enveloped RNA virus found diversely in human and wildlife. WHO has declared this outbreak as public health emergency of international concern. Almost 14,761,814 cases have been identified across the world as on 19 July 2020. The damage mechanism of disease is still unclear. This review will introduce and systematically summarizes the origin and spread of disease along with its epidemiology, pathogenesis, clinical characters with different stages, diagnosis and treatment strategies. This is expected that this review will be helpful for the public in recognising and deal with threatening disease and can also provide a reference for further studies. It will also provide a means to raise awareness about the pandemic among the public and healthcare